refactor: delete user-persona domain (#33)
