--- src/main/java/com/kt/mindLog/domain/user/UserPersona.java
@@ -1,34 +1,0 @@
-package com.kt.mindLog.domain.user;
-
-import java.util.UUID;
-
-import org.hibernate.annotations.UuidGenerator;
-
-import com.kt.mindLog.domain.persona.Persona;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-
-@Getter
-@Entity
-@Table(name = "user_persona")
-public class UserPersona {
-	@Id
-	@UuidGenerator(style = UuidGenerator.Style.VERSION_7)
-	@Column(name = "user_persona_id")
-	private UUID id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "persona_id", nullable = false)
-	private Persona persona;
-}
