+++ src/main/java/com/kt/mindLog/service/persona/PersonaService.java
@@ -11,8 +11,7 @@ import com.kt.mindLog.domain.user.Role;
 import com.kt.mindLog.dto.persona.request.PersonaCreateRequest;
 import com.kt.mindLog.global.common.exception.ErrorCode;
 import com.kt.mindLog.global.common.support.Preconditions;
-import com.kt.mindLog.repository.persona.PersonaRepository;
-import com.kt.mindLog.repository.persona.UserPersonaRepository;
+import com.kt.mindLog.repository.PersonaRepository;
 import com.kt.mindLog.service.s3.S3Path;
 import com.kt.mindLog.service.s3.S3Service;
 
@@ -25,7 +24,6 @@ import tools.jackson.databind.ObjectMapper;
 @RequiredArgsConstructor
 public class PersonaService {
 	private final PersonaRepository personaRepository;
-	private final UserPersonaRepository userPersonaRepository;
 
 	private final S3Service s3Service;
 	private final ObjectMapper objectMapper;
